# Faster fastpath