@max-sixty 2024-05-05
Merging pending tests. I'm not 100% on the implications of skipping `_maybe_wrap_data` in some cases, so if anyone knows whether this was required, we can reassess. Thanks @hmaarrfk